requirements.txt
Flask==1.1.2
flask-restplus==0.13.0
google-cloud-ndb==1.7.1
Werkzeug<1.0.0
pyrsistent==0.16.0
more-itertools==5.0.0
setuptools==40.8.0
timezonefinder==4.5.0